service for retrieve tickets

+++ internal/queries/ticket_status_queries.go
@@ -14,7 +14,7 @@ func NewTicketStatusQueries(db *gorm.DB) *TicketStatusQueries {
 	return &TicketStatusQueries{db: db}
 }
 
-func (q *TicketStatusQueries) FindTicketStatusByID(statusID uint) (*modeldb.TicketStatus, error) {
+func (q *TicketStatusQueries) FindTicketStatusByID(statusID int64) (*modeldb.TicketStatus, error) {
 	var status modeldb.TicketStatus
 	if err := q.db.First(&status, statusID).Error; err != nil {
 		return nil, err
